+++ docs/releases/releases.md
@@ -18,6 +18,8 @@ The main changes in this release are:
 
 + Added `bayesian_search_forecaster_multiseries` function to `model_selection_multiseries` module. This function performs a Bayesian hyperparameter search for the `ForecasterAutoregMultiSeries`, `ForecasterAutoregMultiSeriesCustom`, and `ForecasterAutoregMultiVariate` using `optuna` as the search engine.
 
++ Added argument `output_file` to `bayesian_search_forecaster` function in the `model_selection` module to save the results of the hyperparameter search to a .csv file.
+
 + Added `save_custom_functions` argument to the `save_forecaster` function in the `utils` module. If `True`, save custom functions used in the forecaster (`fun_predictors` and `weight_func`) as .py files. Custom functions must be available in the environment where the forecaster is loaded.
 
 + Added `select_features` function to the `model_selection` module to perform feature selection using scikit-learn selectors.
@@ -32,7 +34,7 @@ The main changes in this release are:
 
 **Changed**
 
-+ 
++  Deprecated argument `lags_grid` in `bayesian_search_forecaster`. Use `search_space` to define the candidate values for the lags. This way, lags can be optimized together with the other parameters of the regressor in the bayesian search. 
 
 + Changed the default value of the `transformer_series` argument to use a `StandardScaler()` in the Global Forecasters (`ForecasterAutoregMultiSeries`, `ForecasterAutoregMultiSeriesCustom` and `ForecasterAutoregMultiVariate`).